Fix CI: ruff RUF059/RUF005/RUF046 lint errors and mypy type annotation
Upgraded ruff 0.9→0.15 caught 16 new lint violations (unused unpacked
vars, list concat style, redundant int cast). Also fixed mypy
no-any-return in global_registry.

+++ lib/resources.py
@@ -151,7 +151,7 @@ def check_cpu() -> CPUInfo:
                 m = re.search(r"([\d.]+)%\s*idle", line)
                 if m:
                     idle = float(m.group(1))
-                    usage = int(round(100 - idle))
+                    usage = round(100 - idle)
                 break
 
     if usage >= CPU_SATURATED:
